Rename config modules to drop _config.
