+++ changelog.d/9816.misc
@@ -1,0 +1,1 @@
+Rename some handlers and config modules to not duplicate the top-level module.
