+++ synapse/config/consent.py
