+++ synapse/config/server_notices.py
